% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_buildphylotree.R
\name{build_phylo_tree}
\alias{build_phylo_tree}
\title{Step 5 - Build Phylogenetic Tree}
\usage{
build_phylo_tree()
}
\description{
The phangorn R package is then used to construct a phylogenetic
tree. First, a neighbor-joining tree is constructed and then fit to a GTR+G+I
maximum likelihood tree using the neighbor-joining tree as a starting point.
}
