% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_buildphyloseq.R
\name{build_phyloseq}
\alias{build_phyloseq}
\title{Step 6 - Build Phyloseq Object}
\usage{
build_phyloseq()
}
\description{
The phyloseq package organizes and synthesizes the different data
types from a typical amplicon sequencing experiment into a single data object
that can be easily manipulated.
}
