% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_buildqp.R
\name{build_plot}
\alias{build_plot}
\title{Step 1 - Build Quality Plot}
\usage{
build_plot(i)
}
\arguments{
\item{i}{is number of reads to be sampled from the available data sets.}
}
\description{
Forward and Reverse Error Profiles, the mean is in green, the median the
solid orange line and the quartiles are the dotted orange lines.
}
\examples{
#' build_plot(i = 3)

}
