% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_buildseqtable.R
\name{sequence_table}
\alias{sequence_table}
\title{Step 4 - Construct Sequence Table and Remove Chimeras}
\usage{
sequence_table(ref_fasta)
}
\arguments{
\item{ref_fasta}{This is the reference file used for assigning genomic data.}
}
\description{
Chimeric sequences are removed by comparing each inferred sequence
to the others in the table, and removing those that can be reproduced by stitching
together two more abundant sequences.
}
