% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_truncate.R
\name{Trunc}
\alias{Trunc}
\title{Step 2 - Truncate Reads}
\usage{
Trunc(L, f, r)
}
\arguments{
\item{L}{number of nucleotides to trim from the left position}

\item{f}{truncate value for forward reads}

\item{r}{truncate value for reverse reads}
}
\description{
Truncate forward and reverse reads based on quality plot output.
}
\examples{
Trunc(L = 10, f = 245, r = 160)

}
